Move CecilExtensions and TestPackageExtensions to Extensions namespace
