+++ src/TestEngine/nunit.engine.core/Runners/DirectTestRunner.cs
@@ -1,4 +1,4 @@
-﻿// ***********************************************************************
+// ***********************************************************************
 // Copyright (c) 2011-2014 Charlie Poole, Rob Prouse
 //
 // Permission is hereby granted, free of charge, to any person obtaining
@@ -25,6 +25,7 @@ using System;
 using System.Collections.Generic;
 using System.IO;
 using NUnit.Engine.Extensibility;
+using NUnit.Engine.Extensions;
 using NUnit.Engine.Internal;
 
 namespace NUnit.Engine.Runners
